+++ docs/api/admin/code_samples/JavaScript/admin_order-edits_{id}/post.js
@@ -1,7 +1,7 @@
 import Medusa from "@medusajs/medusa-js"
 const medusa = new Medusa({ baseUrl: MEDUSA_BACKEND_URL, maxRetries: 3 })
 // must be previously logged in or use api token
-medusa.admin.orderEdits.update(order_edit_id, {
+medusa.admin.orderEdits.update(orderEditId, {
   internal_note: "internal reason XY"
 })
   .then(({ order_edit }) => {