+++ docs/api/admin/code_samples/JavaScript/admin_tax-rates_{id}_products_batch/post.js
@@ -1,9 +1,9 @@
 import Medusa from "@medusajs/medusa-js"
 const medusa = new Medusa({ baseUrl: MEDUSA_BACKEND_URL, maxRetries: 3 })
 // must be previously logged in or use api token
-medusa.admin.taxRates.addProducts(tax_rate_id, {
+medusa.admin.taxRates.addProducts(taxRateId, {
   products: [
-    product_id
+    productId
   ]
 })
 .then(({ tax_rate }) => {