+++ docs/api/admin/code_samples/Shell/admin_auth/post.sh
@@ -1,5 +1,5 @@
-curl --location --request POST 'https://medusa-url.com/admin/auth' \
---header 'Content-Type: application/json' \
+curl -X POST 'https://medusa-url.com/admin/auth' \
+-H 'Content-Type: application/json' \
 --data-raw '{
   "email": "user@example.com",
   "password": "supersecret"