+++ docs/api/admin/code_samples/Shell/admin_customers/post.sh
@@ -1,6 +1,6 @@
-curl --location --request POST 'https://medusa-url.com/admin/customers' \
---header 'Authorization: Bearer {api_token}' \
---header 'Content-Type: application/json' \
+curl -X POST 'https://medusa-url.com/admin/customers' \
+-H 'Authorization: Bearer {api_token}' \
+-H 'Content-Type: application/json' \
 --data-raw '{
     "email": "user@example.com",
     "first_name": "Caterina",