+++ docs/api/admin/code_samples/Shell/admin_orders_{id}_shipment/post.sh
@@ -1,6 +1,6 @@
-curl --location --request POST 'https://medusa-url.com/admin/orders/{id}/shipment' \
---header 'Authorization: Bearer {api_token}' \
---header 'Content-Type: application/json' \
+curl -X POST 'https://medusa-url.com/admin/orders/{id}/shipment' \
+-H 'Authorization: Bearer {api_token}' \
+-H 'Content-Type: application/json' \
 --data-raw '{
     "fulfillment_id": "{fulfillment_id}"
 }'