+++ docs/api/admin/code_samples/Shell/admin_regions_{id}_payment-providers_{provider_id}/delete.sh
@@ -1,2 +1,2 @@
-curl --location --request DELETE 'https://medusa-url.com/admin/regions/{id}/payment-providers/manual' \
---header 'Authorization: Bearer {api_token}'
+curl -X DELETE 'https://medusa-url.com/admin/regions/{id}/payment-providers/{provider_id}' \
+-H 'Authorization: Bearer {api_token}'