+++ docs/api/admin/code_samples/Shell/admin_tax-rates/post.sh
@@ -1,6 +1,6 @@
-curl --location --request POST 'https://medusa-url.com/admin/tax-rates' \
---header 'Authorization: Bearer {api_token}' \
---header 'Content-Type: application/json' \
+curl -X POST 'https://medusa-url.com/admin/tax-rates' \
+-H 'Authorization: Bearer {api_token}' \
+-H 'Content-Type: application/json' \
 --data-raw '{
     "code": "TEST",
     "name": "New Tax Rate",