+++ docs/api/admin/code_samples/Shell/admin_users_reset-password/post.sh
@@ -1,6 +1,6 @@
-curl --location --request POST 'https://medusa-url.com/admin/users/reset-password' \
---header 'Authorization: Bearer {api_token}' \
---header 'Content-Type: application/json' \
+curl -X POST 'https://medusa-url.com/admin/users/reset-password' \
+-H 'Authorization: Bearer {api_token}' \
+-H 'Content-Type: application/json' \
 --data-raw '{
     "token": "supersecrettoken",
     "password": "supersecret"