+++ docs/api/store/code_samples/Shell/store_carts/post.sh
@@ -1,1 +1,1 @@
-curl --location --request POST 'https://medusa-url.com/store/carts'
+curl -X POST 'https://medusa-url.com/store/carts'