+++ integration-tests/plugins/medusa-config.js
@@ -1,3 +1,5 @@
+const { Modules } = require("@medusajs/modules-sdk")
+const { Workflows } = require("@medusajs/workflows")
 const DB_HOST = process.env.DB_HOST
 const DB_USERNAME = process.env.DB_USERNAME
 const DB_PASSWORD = process.env.DB_PASSWORD
@@ -30,22 +32,28 @@ module.exports = {
     cookie_secret: "test",
     database_extra: { idle_in_transaction_session_timeout: 0 },
   },
+  featureFlags: {
+    workflows: {
+      [Workflows.CreateProducts]: true,
+      [Workflows.CreateCart]: true,
+    },
+  },
   modules: {
-    stockLocationService: {
+    [Modules.STOCK_LOCATION]: {
       scope: "internal",
       resources: "shared",
       resolve: "@medusajs/stock-location",
     },
-    inventoryService: {
+    [Modules.INVENTORY]: {
       scope: "internal",
       resources: "shared",
       resolve: "@medusajs/inventory",
     },
-    cacheService: {
+    [Modules.CACHE]: {
       resolve: "@medusajs/cache-inmemory",
       options: { ttl: 5 },
     },
-    productModuleService: {
+    [Modules.PRODUCT]: {
       scope: "internal",
       resources: "shared",
       resolve: "@medusajs/product",