+++ packages/product/src/loaders/container.ts
@@ -1,5 +1,4 @@
 import * as DefaultRepositories from "@repositories"
-
 import {
   BaseRepository,
   ProductCategoryRepository,