+++ packages/product/src/repositories/product-image.ts
@@ -5,11 +5,14 @@ import {
 } from "@mikro-orm/core"
 import { Context, DAL } from "@medusajs/types"
 import { Image, Product } from "@models"
-import { AbstractBaseRepository } from "./base"
 import { SqlEntityManager } from "@mikro-orm/postgresql"
-import { InjectTransactionManager, MedusaContext } from "@medusajs/utils"
+import {
+  DALUtils,
+  InjectTransactionManager,
+  MedusaContext,
+} from "@medusajs/utils"
 
-export class ProductImageRepository extends AbstractBaseRepository<Image> {
+export class ProductImageRepository extends DALUtils.MikroOrmAbstractBaseRepository<Image> {
   protected readonly manager_: SqlEntityManager
 
   constructor({ manager }: { manager: SqlEntityManager }) {