+++ packages/workflows/src/handlers/middlewares/index.ts
@@ -1,0 +1,2 @@
+export * from "./create-products-prepare-create-prices-compensation"
+export * from "./create-products-prepare-data"