+++ packages/workflows/src/helper/pipe.ts
@@ -1,42 +1,57 @@
-import { Context, MedusaContainer, SharedContext } from "@medusajs/types"
 import {
   TransactionMetadata,
   WorkflowStepHandler,
 } from "@medusajs/orchestration"
+import { Context, MedusaContainer, SharedContext } from "@medusajs/types"
 
+import { DistributedTransaction } from "@medusajs/orchestration"
 import { InputAlias } from "../definitions"
 
-type WorkflowStepReturn = {
-  alias: string
+export type WorkflowStepMiddlewareReturn = {
+  alias?: string
   value: any
 }
 
-type WorkflowStepInput = {
+export type WorkflowStepMiddlewareInput = {
   from: string
-  alias: string
+  alias?: string
 }
 
 interface PipelineInput {
   inputAlias?: InputAlias | string
-  invoke?: WorkflowStepInput | WorkflowStepInput[]
-  compensate?: WorkflowStepInput | WorkflowStepInput[]
+  invoke?: WorkflowStepMiddlewareInput | WorkflowStepMiddlewareInput[]
+  compensate?: WorkflowStepMiddlewareInput | WorkflowStepMiddlewareInput[]
+  onComplete?: (args: WorkflowOnCompleteArguments) => {}
 }
 
-export type WorkflowArguments = {
+export type WorkflowArguments<T = any> = {
   container: MedusaContainer
   payload: unknown
-  data: any
+  data: T
   metadata: TransactionMetadata
   context: Context | SharedContext
 }
 
-export type PipelineHandler = (
-  args: WorkflowArguments
-) => Promise<WorkflowStepReturn | WorkflowStepReturn[]>
+export type WorkflowOnCompleteArguments<T = any> = {
+  container: MedusaContainer
+  payload: unknown
+  data: T
+  metadata: TransactionMetadata
+  transaction: DistributedTransaction
+  context: Context | SharedContext
+}
 
-export function pipe(
+export type PipelineHandler<T extends any = undefined> = (
+  args: WorkflowArguments
+) => Promise<
+  T extends undefined
+    ? WorkflowStepMiddlewareReturn | WorkflowStepMiddlewareReturn[]
+    : T
+>
+
+export function pipe<T>(
   input: PipelineInput,
-  ...functions: PipelineHandler[]
+  ...functions: [...PipelineHandler[], PipelineHandler<T>]
 ): WorkflowStepHandler {
   return async ({
     container,
@@ -44,9 +59,10 @@ export function pipe(
     invoke,
     compensate,
     metadata,
+    transaction,
     context,
   }) => {
-    const data = {}
+    let data = {}
 
     const original = {
       invoke: invoke ?? {},
@@ -57,7 +73,8 @@ export function pipe(
       Object.assign(original.invoke, { [input.inputAlias]: payload })
     }
 
-    for (const key in input) {
+    const dataKeys = ["invoke", "compensate"]
+    for (const key of dataKeys) {
       if (!input[key]) {
         continue
       }
@@ -67,13 +84,16 @@ export function pipe(
       }
 
       for (const action of input[key]) {
-        if (action?.alias) {
+        if (action.alias) {
           data[action.alias] = original[key][action.from]
+        } else {
+          data[action.from] = original[key][action.from]
         }
       }
     }
 
-    return functions.reduce(async (_, fn) => {
+    let finalResult
+    for (const fn of functions) {
       let result = await fn({
         container,
         payload,
@@ -88,11 +108,34 @@ export function pipe(
             data[action.alias] = action.value
           }
         }
-      } else if (result?.alias) {
-        data[result.alias] = result.value
+      } else if (
+        result &&
+        "alias" in (result as WorkflowStepMiddlewareReturn)
+      ) {
+        if ((result as WorkflowStepMiddlewareReturn).alias) {
+          data[(result as WorkflowStepMiddlewareReturn).alias!] = (
+            result as WorkflowStepMiddlewareReturn
+          ).value
+        } else {
+          data = (result as WorkflowStepMiddlewareReturn).value
+        }
       }
 
-      return result
-    }, {})
+      finalResult = result
+    }
+
+    if (typeof input.onComplete === "function") {
+      const dataCopy = JSON.parse(JSON.stringify(data))
+      await input.onComplete({
+        container,
+        payload,
+        data: dataCopy,
+        metadata,
+        transaction,
+        context: context as Context,
+      })
+    }
+
+    return finalResult
   }
 }